+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/Decorators/EntityFrameworkCoreRepositoryDecoratorRegistration.cs
@@ -5,9 +5,6 @@ using Intent.Modules.EntityFrameworkCore.Templates.DbContext;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
-[assembly: DefaultIntentManaged(Mode.Fully)]
-[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.TemplateDecoratorRegistration", Version = "1.0")]
-
 namespace Intent.Modules.EntityFrameworkCore.Repositories.Decorators
 {
     [Description(EntityFrameworkCoreRepositoryDecorator.DecoratorId)]