+++ Modules/Intent.Modules.Infrastructure.DependencyInjection/Templates/DependencyInjection/DependencyInjectionTemplatePartial.cs
@@ -87,9 +87,21 @@ namespace Intent.Modules.Infrastructure.DependencyInjection.Templates.Dependency
 
             @event.MarkAsHandled();
             _containerRegistrationRequests.Add(@event);
+            
             foreach (var templateDependency in @event.TemplateDependencies)
             {
+                var template = GetTemplate<IClassProvider>(templateDependency);
+                if (template != null)
+                {
+                    AddUsing(template.Namespace);
+                }
+                
                 AddTemplateDependency(templateDependency);
+            }
+
+            foreach (var ns in @event.RequiredNamespaces)
+            {
+                AddUsing(ns);
             }
         }
 
@@ -102,6 +114,7 @@ namespace Intent.Modules.Infrastructure.DependencyInjection.Templates.Dependency
 
             @event.MarkAsHandled();
             _serviceConfigurationRequests.Add(@event);
+            
             foreach (var templateDependency in @event.TemplateDependencies)
             {
                 var template = GetTemplate<IClassProvider>(templateDependency);
@@ -111,6 +124,11 @@ namespace Intent.Modules.Infrastructure.DependencyInjection.Templates.Dependency
                 }
 
                 AddTemplateDependency(templateDependency);
+            }
+            
+            foreach (var ns in @event.RequiredNamespaces)
+            {
+                AddUsing(ns);
             }
         }