+++ Modules/Intent.Modules.MongoDb.Repositories/FactoryExtensions/MongoRepoFactoryExtension.cs
@@ -2,15 +2,10 @@ using System.Linq;
 using Intent.Engine;
 using Intent.Modelers.Domain.Api;
 using Intent.Modules.Common;
-using Intent.Modules.Common.CSharp.DependencyInjection;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Plugins;
 using Intent.Modules.Common.Templates;
-using Intent.Modules.Entities.Repositories.Api.Templates;
 using Intent.Modules.Entities.Repositories.Api.Templates.EntityRepositoryInterface;
-using Intent.Modules.Entities.Repositories.Api.Templates.RepositoryInterface;
-using Intent.Modules.Entities.Repositories.Api.Templates.UnitOfWorkInterface;
-using Intent.Modules.MongoDb.Templates.ApplicationMongoDbContext;
 using Intent.MongoDb.Api;
 using Intent.Plugins.FactoryExtensions;
 using Intent.RoslynWeaver.Attributes;
@@ -38,10 +33,9 @@ namespace Intent.Modules.MongoDb.Repositories.FactoryExtensions
         /// </remarks>
         protected override void OnBeforeTemplateExecution(IApplication application)
         {
-            UpdateDbContextTemplate(application);
             UpdateRepositoryInterfaceTemplate(application);
         }
-
+        
         private static void UpdateRepositoryInterfaceTemplate(IApplication application)
         {
             var repositoryTemplates = application.FindTemplateInstances<ICSharpFileBuilderTemplate>(TemplateDependency.OnTemplate(EntityRepositoryInterfaceTemplate.TemplateId));
@@ -63,32 +57,6 @@ namespace Intent.Modules.MongoDb.Repositories.FactoryExtensions
                         .AddParameter(repositoryTemplate.GetTypeName("Domain.Entity.Interface", model), "entity");
                 });
             }
-        }
-
-        private static void UpdateDbContextTemplate(IApplication application)
-        {
-            var dbContextTemplate = application.FindTemplateInstance<ICSharpFileBuilderTemplate>(TemplateDependency.OnTemplate(ApplicationMongoDbContextTemplate.TemplateId));
-            if (dbContextTemplate == null)
-            {
-                return;
-            }
-
-            var intentTemplate = (CSharpTemplateBase<object>)dbContextTemplate;
-            var @class = dbContextTemplate.CSharpFile.Classes.First();
-            @class.ImplementsInterface(intentTemplate.GetUnitOfWorkInterfaceName());
-            @class.AddMethod("Task<int>", "IUnitOfWork.SaveChangesAsync", method =>
-            {
-                method.Async();
-                method.WithoutAccessModifier();
-                method.AddParameter("CancellationToken", "cancellationToken");
-                method.AddStatement("return (await base.SaveChangesAsync(cancellationToken)).Results.Count;");
-            });
-
-            dbContextTemplate.ExecutionContext.EventDispatcher.Publish(ContainerRegistrationRequest
-                .ToRegister(intentTemplate)
-                .ForInterface(intentTemplate.GetTemplate<IClassProvider>(UnitOfWorkInterfaceTemplate.TemplateId))
-                .ForConcern("Infrastructure")
-                .WithResolveFromContainer());
         }
     }
 }