+++ Modules/Intent.Modules.Application.DependencyInjection/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.3.10
+
+- Fixed: Didn't take "requires namespaces" into account and logic from Infrastructure around Container registration has been aligned. 
+
 ### Version 3.3.8
 
 - Update: Dependency injection configuration now receives event updates from `ServiceConfigurationRequest` and not just from `ContainerRegistrationRequest`.